Revert "Dev to Master, v1.4 update (#1)"
This reverts commit 9c897949ee8bc71e886d2c85283a4dd94fb3420b.
