+++ ScanThoseGates.version
@@ -5,9 +5,9 @@
 	"modVersion":
 	{
 		"major":1,
-		"minor":4,
+		"minor":3,
 		#"patch":,
 	},
 	"starsectorVersion":"0.95.1a-RC6",
-	"directDownloadURL":"https://github.com/AEROassault/ScanThoseGates/releases/download/v1.3/ScanThoseGates.zip",
+	"directDownloadURL":"https://github.com/AEROassault/ScanThoseGates/releases/download/1.3/ScanThoseGates.zip",
 }
