+++ mod_info.json
@@ -2,8 +2,8 @@
   "id":"scan_those_gates",
   "name":"Scan Those Gates",
   "author":"AERO",
-  "version":"1.4",
-  "description":"Scans all gates that are located in systems with at least one non-hidden market at save start",
+  "version":"1.3",
+  "description":"Scans all gates that do not have 'Inactive' in their name at the beginning of a save when Nexerelin Skip Story is enabled",
   "gameVersion":"0.95.1a-RC6",
   "jars":["jars/ScanThoseGates.jar"],
   "modPlugin":"data.scripts.stg_ModPlugin"
